Moved ReviewListAdapter to adapter package
